+++ templates/encrypt-password.html
@@ -14,7 +14,6 @@
 					<a class="nav-link" href="/">HOME</a>
 					<a class="current-page">ENCRPYT</a>
 					<a class="nav-link" href="/view-list">VIEW</a>
-					<menu class="navbar-nav">
 				</div>
 			</nav>
 			<a href="/">
