+++ templates/index.html
@@ -14,7 +14,6 @@
 					<a class="current-page">HOME</a>
 					<a class="nav-link" href="/encrypt-password">ENCRPYT</a>
 					<a class="nav-link" href="/view-list">VIEW</a>
-					<menu class="navbar-nav">
 				</div>
 			</nav>
 	        <h1>TriVon Password Manager (Alpha Testing)</h1>
