+++ templates/view-list.html
@@ -14,13 +14,12 @@
 					<a class="nav-link" href="/">HOME</a>
 					<a class="nav-link" href="/encrypt-password">ENCRPYT</a>
 					<a class="current-page">VIEW</a>
-					<menu class="navbar-nav">
 				</div>
 			</nav>
 	        <h2>Account List:</h2>
 	        {% for accounts in account_list %}
 	            <h4>{{accounts.account}}:</h4>
-	            <p>{{accounts.password}}</p>
+	            <p>{{accounts.passwordManager.decrypt_password()}}</p>
 	            <hr style="width: 25%; margin-left:0">
                 {% endfor %}
 	    </body>
